+++ src/Activity/activityProvider.js
@@ -19,27 +19,7 @@ const activityProvider = {
         } catch(err) {
             return {error: true};
         }
-    }, 
-
-    retrieveActivitiesByPlaceId: async (placeId) => {
-        try {
-            const connection = await pool.getConnection(async conn => conn);
-            const activitiesResult = await activityDao.selectActivitiesByPlaceId(connection, placeId);
-            console.log(activitiesResult);
-            console.log(typeof(activitiesResult));
-            if(Object.keys(activitiesResult).length == 0) {
-                return {error: true};
-            }
-            connection.release();
-            return {
-                error: false,
-                result: activitiesResult
-            };
-        } catch (err) {
-            return {error: true};
-        }
-    }
-
+    },
 }
 
 export default activityProvider;
