+++ src/Activity/activityRouter.js
@@ -1,0 +1,9 @@
+import express from "express";
+import activityController from "./activityController";
+
+
+const activityRouter = express.Router();
+
+activityRouter.get('/:activityId(\\d+)', activityController.getActivityByActivityId);
+
+export default activityRouter;
