+++ src/places/placesController.js
@@ -99,6 +99,21 @@ const placesController = {
       return res.status(500).json(response(baseResponse.SERVER_ERROR));
     }
   },
+  getActivities: async (req, res) => {
+    try {
+      const placeId = req.params.placeId;
+
+      const activities = await placesService.retrieveActivitiesByPlaceId(placeId);
+
+      if(activities.error) {
+        return res.status(400).json(response(baseResponse.LOCATION_EMPTY));
+      }
+      return res.status(200).json(response(baseResponse.SUCCESS, activities))
+    } catch (error) {
+      //console.log(error);
+      return res.status(500).json(response(baseResponse.SERVER_ERROR));
+    }
+  }
 };
 
 export default placesController;
