+++ src/places/placesRouter.js
@@ -8,5 +8,5 @@ placesRouter.get("/", placesController.getPlaces);
 placesRouter.get("/search", placesController.searchPlaces);
 placesRouter.get("/:placeId(\\d+)/reviews", placesController.getReviews);
 placesRouter.post("/:placeId(\\d+)/reviews", placesController.postReview);
-
+placesRouter.get("/:placeId(\\d+)/activities", placesController.getActivities);
 export default placesRouter;
