+++ src/places/placesService.js
@@ -139,6 +139,24 @@ const placesService = {
 
     return searchResult;
   },
+  retrieveActivitiesByPlaceId: async (placeId) => {
+    try {
+      const connection = await pool.getConnection(async conn => conn);
+      const activitiesResult = await placesDao.selectActivitiesByPlaceId(connection, placeId);
+      console.log(activitiesResult);
+      console.log(typeof(activitiesResult));
+      if(Object.keys(activitiesResult).length == 0) {
+        return {error: true};
+      }
+      connection.release();
+      return {
+        error: false,
+        result: activitiesResult
+      };
+    } catch (err) {
+      return {error: true};
+    }
+  }
 };
 
 export default placesService;
